+++ src/components/Dao/DaoContainer.tsx
@@ -1,6 +1,6 @@
-import { IDAOState, Member } from "@daostack/arc.js";
+import { IDAOState, Member, Scheme } from "@daostack/arc.js";
 import { getProfilesForAddresses } from "actions/profilesActions";
-import CreateProposalPage from "components/Proposal/Create/CreateProposalPage";
+import CreateProposalPage from "components/Proposal/Create";
 import ProposalDetailsPage from "components/Proposal/ProposalDetailsPage";
 import SchemeContainer from "components/Scheme/SchemeContainer";
 import Loading from "components/Shared/Loading";
@@ -17,13 +17,13 @@ import { IProfileState } from "reducers/profilesReducer";
 import DetailsPageRouter from "components/Scheme/ContributionRewardExtRewarders/DetailsPageRouter";
 import { combineLatest, Subscription } from "rxjs";
 import DaoSchemesPage from "./DaoSchemesPage";
-import DaoHistoryPage from "./DaoHistoryPage";
 import DaoMembersPage from "./DaoMembersPage";
 import * as css from "./Dao.scss";
-import DaoLandingPage from "components/Dao/DaoLandingPage";
-import { standardPolling, targetedNetwork, getArcByDAOAddress, getDAONameByID } from "lib/util";
+import DaoProposalsPage from "components/Dao/DaoProposalsPage";
+import { standardPolling, getArcByDAOAddress, getDAONameByID, getNetworkByDAOAddress } from "lib/util";
 import gql from "graphql-tag";
-import { getArcs } from "arc";
+import { enableWalletProvider, getArcs } from "arc";
+import { getKnownSchemes } from "lib/schemeUtils";
 
 type IExternalProps = RouteComponentProps<any>;
 
@@ -34,8 +34,17 @@ interface IStateProps {
   followingDaosAddresses: Array<any>;
 }
 
+enum PROPOSAL_WINDOW_TYPE {
+  PROPOSAL_LIST = "PROPOSAL_LIST",
+  SPECIFIC_PLUGIN = "SPECIFIC_PLUGIN",
+}
+
 interface IState {
   memberDaos: Array<any>;
+  modalParentPath: string;
+  schemeId: string|null;
+  proposalWindowType: PROPOSAL_WINDOW_TYPE|null;
+  loading: boolean;
 }
 
 interface IDispatchProps {
@@ -43,7 +52,7 @@ interface IDispatchProps {
   showNotification: typeof showNotification;
 }
 
-type IProps = IExternalProps & IStateProps & IDispatchProps & ISubscriptionProps<[IDAOState, Member[]]>;
+type IProps = IExternalProps & IStateProps & IDispatchProps & ISubscriptionProps<[IDAOState, Member[], Scheme[]]>;
 
 const mapStateToProps = (state: IRootState, ownProps: IExternalProps): IExternalProps & IStateProps => {
   return {
@@ -65,6 +74,10 @@ class DaoContainer extends React.Component<IProps, IState> {
     super(props);
     this.state = {
       memberDaos: [],
+      modalParentPath: "/dao/:daoAvatarAddress",
+      schemeId: "",
+      proposalWindowType: null,
+      loading: false,
     };
   }
   public daoSubscription: any;
@@ -99,7 +112,45 @@ class DaoContainer extends React.Component<IProps, IState> {
     this.setState({ memberDaos: memberDaos });
   }
 
-  private daoHistoryRoute = (routeProps: any) => <DaoHistoryPage {...routeProps} daoState={this.props.data[0]} currentAccountAddress={this.props.currentAccountAddress} />;
+  componentDidUpdate(prevProps: IProps) {
+    if (this.props.data[0] !== prevProps.data[0]){
+      this.setState({ loading: false });
+    }
+  }
+
+  public handleNewProposal = async (schemeId?: string): Promise<void> => {
+    const { showNotification, daoAvatarAddress } = this.props;
+
+    if (!await enableWalletProvider({ showNotification }, getNetworkByDAOAddress(daoAvatarAddress))) { return; }
+
+    if (typeof schemeId === "string") {
+      this.setState(() => ({
+        modalParentPath: "/dao/:daoAvatarAddress/scheme/:schemeId",
+        schemeId,
+        proposalWindowType: PROPOSAL_WINDOW_TYPE.SPECIFIC_PLUGIN,
+      }), () => {
+        this.props.history.push(`/dao/${daoAvatarAddress}/scheme/${schemeId}/proposals/create/`);
+      });
+    } else {
+      this.setState(() => ({
+        modalParentPath: "/dao/:daoAvatarAddress",
+        schemeId: "",
+        proposalWindowType: PROPOSAL_WINDOW_TYPE.PROPOSAL_LIST,
+      }), () => {
+        this.props.history.push(`/dao/${daoAvatarAddress}/proposals/create`);
+      });
+    }
+  };
+
+  public handleProposalWindowClose = (parentPath: string) => {
+    const { history } = this.props;
+    history.push(parentPath);
+    this.setState({
+      schemeId: "",
+      proposalWindowType: null,
+    });
+  }
+
   private daoMembersRoute = (routeProps: any) => <DaoMembersPage {...routeProps} daoState={this.props.data[0]} />;
   private daoProposalRoute = (routeProps: any) =>
     <ProposalDetailsPage {...routeProps}
@@ -114,17 +165,46 @@ class DaoContainer extends React.Component<IProps, IState> {
       proposalId={routeProps.match.params.proposalId}
     />;
 
-  private schemeRoute = (routeProps: any) => <SchemeContainer {...routeProps} daoState={this.props.data[0]} currentAccountAddress={this.props.currentAccountAddress} />;
+  private schemeRoute = (routeProps: any) => (
+    <SchemeContainer
+      {...routeProps}
+      daoState={this.props.data[0]}
+      currentAccountAddress={this.props.currentAccountAddress}
+      onCreateProposal={this.handleNewProposal}
+    />
+  );
   private daoSchemesRoute = (routeProps: any) => <DaoSchemesPage {...routeProps} daoState={this.props.data[0]} />;
-  private daoLandingRoute = (_routeProps: any) => <DaoLandingPage daoState={this.props.data[0]} />;
-  private modalRoute = (route: any) => `/dao/${route.params.daoAvatarAddress}/scheme/${route.params.schemeId}/`;
-  private onFollwingDaosListChange = (daoAddress: string, history: any) => history.push(`/dao/${daoAddress}`);
+  private daoProposalsRoute = (routeProps: any) => (
+    <DaoProposalsPage
+      {...routeProps}
+      daoState={this.props.data[0]}
+      schemesLength={getKnownSchemes(this.props.data[2]).length}
+      currentAccountAddress={this.props.currentAccountAddress}
+      onCreateProposal={this.handleNewProposal}
+    />
+  );
+  private createProposalRoute = (routeProps: any) => (
+    <CreateProposalPage
+      {...routeProps}
+      onBackToParent={this.handleProposalWindowClose}
+      dao={this.props.data[0].dao}
+      currentAccountAddress={this.props.currentAccountAddress}
+    />
+  );
+  private modalRouteCreateProposal = (route: any) => {
+    return this.state.modalParentPath
+      .replace(":daoAvatarAddress", route.params.daoAvatarAddress)
+      .replace(":schemeId", this.state.schemeId);
+  };
+  private onFollwingDaosListChange = (daoAddress: string, history: any) => {
+    this.setState({ loading: true });
+    history.push(`/dao/${daoAddress}`);
+  };
 
   public render(): RenderOutput {
     const daoState = this.props.data[0];
-    const network = targetedNetwork();
     const { followingDaosAddresses } = this.props;
-    const { memberDaos } = this.state;
+    const { memberDaos, loading } = this.state;
 
     const followingDaos = followingDaosAddresses?.filter(address => getDAONameByID(address) !== undefined)
     .map(daoAddress => {
@@ -132,11 +212,11 @@ class DaoContainer extends React.Component<IProps, IState> {
     });
 
     const myDaos = followingDaos?.concat(memberDaos);
-
+    const currentDaoAddress = window.location.pathname.split("/")[2];
     const myDaosAddresses = [] as any;
     const myDaosOptions = myDaos?.map(dao => {
       myDaosAddresses.push(dao.id);
-      return <option key={dao.id} selected={dao.id === daoState.address ? true : false} value={dao.id}>{dao.name}</option>;
+      return <option key={dao.id} selected={dao.id === currentDaoAddress ? true : false} value={dao.id}>{dao.name}</option>;
     });
 
     if (!myDaosAddresses.includes(daoState.id)){
@@ -154,45 +234,41 @@ class DaoContainer extends React.Component<IProps, IState> {
           <meta name="twitter:description" content={daoState.name + " | Managed on Alchemy by DAOstack"} />
         </Helmet>
 
-        <div className={css.wrapper}>
-          <div className={css.noticeWrapper}>
-            <div className={css.noticeBuffer}></div>
-            <div className={css.notice}>Alchemy 2.0 has been released! Take a look <a
-              href={(network === "main") ? process.env.ALCHEMY_V2_URL_MAINNET : process.env.ALCHEMY_V2_URL_XDAI}
-              target="_blank" rel="noopener noreferrer">here</a>.
-            </div>
-          </div>
+        {!loading ? <div className={css.wrapper}>
           <Switch>
-            <Route exact path="/dao/:daoAvatarAddress"
-              render={this.daoLandingRoute} />
-            <Route exact path="/dao/:daoAvatarAddress/history"
-              render={this.daoHistoryRoute} />
             <Route exact path="/dao/:daoAvatarAddress/members"
               render={this.daoMembersRoute} />
 
             <Route exact path="/dao/:daoAvatarAddress/proposal/:proposalId"
-              render={this.daoProposalRoute}
-            />
+              render={this.daoProposalRoute} />
 
             <Route path="/dao/:daoAvatarAddress/crx/proposal/:proposalId"
               render={this.daoCrxProposalRoute} />
 
-            <Route path="/dao/:daoAvatarAddress/scheme/:schemeId"
+            <Route exact={this.state.proposalWindowType === PROPOSAL_WINDOW_TYPE.PROPOSAL_LIST} path="/dao/:daoAvatarAddress/scheme/:schemeId"
               render={this.schemeRoute} />
 
             <Route exact path="/dao/:daoAvatarAddress/schemes"
               render={this.daoSchemesRoute} />
 
-            <Route path="/dao/:daoAvatarAddress" render={this.daoLandingRoute} />
+            <Route path="/dao/:daoAvatarAddress"
+              render={this.daoProposalsRoute}
+            />
           </Switch>
 
           <ModalRoute
-            path="/dao/:daoAvatarAddress/scheme/:schemeId/proposals/create"
-            parentPath={this.modalRoute}
-            component={CreateProposalPage}
+            path="/dao/:daoAvatarAddress/proposals/create"
+            parentPath={this.modalRouteCreateProposal}
+            component={this.createProposalRoute}
           />
 
-        </div>
+          <ModalRoute
+            path="/dao/:daoAvatarAddress/scheme/:schemeId/proposals/create"
+            parentPath={this.modalRouteCreateProposal}
+            component={this.createProposalRoute}
+          />
+
+        </div> : <Loading />}
       </div>
     );
   }
@@ -209,7 +285,8 @@ const SubscribedDaoContainer = withSubscription({
     const dao = arc.dao(daoAddress);
     const observable = combineLatest(
       dao.state(standardPolling(true)), // DAO state
-      dao.members()
+      dao.members(),
+      dao.schemes({ where: { isRegistered: true } }, standardPolling(true)),
     );
     return observable;
   },