+++ src/components/Proposal/Create/SelectProposal/SelectProposalLabel/SelectProposalLabel.tsx
@@ -1,0 +1,20 @@
+import * as React from "react";
+
+import { ISchemeState } from "@daostack/arc.js";
+import { schemeName } from "lib/schemeUtils";
+
+import css from "./SelectProposalLabel.scss";
+
+interface IProps {
+  value: string;
+  label: ISchemeState;
+}
+
+export const SelectProposalLabel: React.FC<IProps> = ({ label, value }) => {
+  return (
+    <div key={value}>
+      <span className={css.label}>{schemeName(label)}</span>
+      <span>({label.address.slice(0, 6)}...{label.address.slice(-4)})</span>
+    </div>
+  );
+};