+++ src/layouts/SidebarMenu.tsx
@@ -122,20 +122,7 @@ class SidebarMenu extends React.Component<IProps, IStateProps> {
         <div className={css.daoNavigation}>
           <ul>
             <li>
-              <Link to={`/dao/${dao.address}`} onClick={this.handleCloseMenu} data-test-id="daohome">
-                <span className={css.menuDot} />
-                <span className={
-                  classNames({
-                    [css.notification]: true,
-                    [css.homeNotification]: true,
-                  })
-                }></span>
-                <img src="/assets/images/Icon/menu/home.svg" />
-                Home
-              </Link>
-            </li>
-            <li>
-              <Link to={`/dao/${dao.address}/schemes`} onClick={this.handleCloseMenu} data-test-id="daoschemes">
+              <Link to={`/dao/${dao.address}`} onClick={this.handleCloseMenu} data-test-id="daoschemes">
                 <span className={css.menuDot} />
                 <span className={
                   classNames({
@@ -145,6 +132,19 @@ class SidebarMenu extends React.Component<IProps, IStateProps> {
                 }></span>
                 <img src="/assets/images/Icon/menu/chat.svg" />
                 Proposals
+              </Link>
+            </li>
+            <li>
+              <Link to={`/dao/${dao.address}/schemes`} onClick={this.handleCloseMenu}>
+                <span className={css.menuDot} />
+                <span className={
+                  classNames({
+                    [css.notification]: true,
+                    [css.historyNotification]: true,
+                  })
+                }></span>
+                <img src="/assets/images/Icon/menu/plugins.svg" />
+                Plugins
               </Link>
             </li>
             <li>
@@ -161,19 +161,6 @@ class SidebarMenu extends React.Component<IProps, IStateProps> {
                   Members
                 </Link>
               </TrainingTooltip>
-            </li>
-            <li>
-              <Link to={"/dao/" + dao.address + "/history/"} onClick={this.handleCloseMenu}>
-                <span className={css.menuDot} />
-                <span className={
-                  classNames({
-                    [css.notification]: true,
-                    [css.historyNotification]: true,
-                  })
-                }></span>
-                <img src="/assets/images/Icon/menu/history.svg" />
-                History
-              </Link>
             </li>
           </ul>
         </div>
@@ -211,8 +198,9 @@ class SidebarMenu extends React.Component<IProps, IStateProps> {
       [css.noDAO]: !this.props.daoAvatarAddress,
       clearfix: true,
     });
+    const daoAddress = this.props.daoAvatarAddress ?? this.props.data?.dao.id;
 
-    const network = targetedNetwork();
+    const network = daoAddress ? getNetworkByDAOAddress(daoAddress) : targetedNetwork();
     const testNet = !((network === "main") || (network === "xdai"));
 
     return (
@@ -251,12 +239,6 @@ class SidebarMenu extends React.Component<IProps, IStateProps> {
                   <li><a className="externalLink"
                     href={(network === "main") ? process.env.ALCHEMY_V2_URL_MAINNET : process.env.ALCHEMY_V2_URL_XDAI}
                     target="_blank" rel="noopener noreferrer">Switch to v2</a></li>
-                  {(network === "main") ?
-                    <li><a className="externalLink" href={process.env.ALCHEMY_V1_URL_XDAI} target="_blank" rel="noopener noreferrer">Switch to xDAI</a></li>
-                    : (network === "xdai") ?
-                      <li><a className="externalLink" href={process.env.ALCHEMY_V1_URL_MAINNET} target="_blank" rel="noopener noreferrer">Switch to Mainnet</a></li>
-                      : ""
-                  }
                 </>
               }
               <li className={css.daoStack}>