+++ test/integration/proposal-schemeregistrar.e2e.ts
@@ -1,9 +1,10 @@
 import * as uuid from "uuid";
-import { getContractAddresses, hideCookieAcceptWindow, gotoDaoSchemes } from "./utils";
+import { getContractAddresses, hideCookieAcceptWindow, gotoDaoSchemes, wait} from "./utils";
 
 describe("SchemeRegistrar Proposals", () => {
   let daoAddress: string;
   let addresses;
+  const GOOD_PARAMS_HASH = "0x7603c16050b9322f37cb156f90baecdd28edbaef67aac17d1164aabf408d35e2";
 
   before(() => {
     addresses = getContractAddresses();
@@ -38,15 +39,18 @@ describe("SchemeRegistrar Proposals", () => {
     await descriptionInput.setValue(`https://this.must.be/a/valid/url${uuid()}`);
 
     const schemeToAddInput = await $("*[id=\"schemeToAddInput\"]");
-    await schemeToAddInput.setValue("0x5fB320886aF629122736c0e1a5c94dCE841EA37B");
+    await schemeToAddInput.setValue("0x4a1d2a5060c782049ef966d9412f1239e95183b7");
 
     const parametersHashInput = await $("*[id=\"parametersHashInput\"]");
-    await parametersHashInput.setValue("0x0000000000000000000000000000000000000000000000000000000000001234");
-
-    // const registerOtherSchemesInput = await $("*[id=\"registerOtherSchemesInput\"]");
-    // await registerOtherSchemesInput.setValue(true);
+    await parametersHashInput.setValue(GOOD_PARAMS_HASH);
 
     const createProposalSubmitButton = await $("*[type=\"submit\"]");
+    /**
+     * The below is a workaround to wait until verifyParametersHash function is finished.
+     * Ideally, we need to disable the submit button and click on it only if there are no errors in the form,
+     * however this raises new issues and should be handled separately.
+     */
+    await wait(1000);
     await createProposalSubmitButton.click();
 
     // check that the proposal appears in the list
@@ -85,7 +89,7 @@ describe("SchemeRegistrar Proposals", () => {
     await schemeToEditInput.selectByIndex(2);
 
     const parametersHashInput = await $("*[id=\"parametersHashInput\"]");
-    await parametersHashInput.setValue("0x0000000000000000000000000000000000000000000000000000000000001234");
+    await parametersHashInput.setValue(GOOD_PARAMS_HASH);
 
     const createProposalSubmitButton = await $("*[type=\"submit\"]");
     await createProposalSubmitButton.click();